Create ProfileMenu component
